odified the code for specific user only can update or delete assignment

+++ index.js
@@ -1,41 +1,41 @@
 const express = require("express");
 const cors = require("cors");
 require("dotenv").config();
-const jwt = require('jsonwebtoken');
-const cookieParser = require('cookie-parser');
+const jwt = require("jsonwebtoken");
+const cookieParser = require("cookie-parser");
 const app = express();
 const port = process.env.PORT || 5000;
 const { MongoClient, ServerApiVersion, ObjectId } = require("mongodb");
 
-
 // middlewares
-app.use(cors({
-    origin: ['http://localhost:5173'],
-    credentials: true
-}));
+app.use(
+  cors({
+    origin: ["http://localhost:5173"],
+    credentials: true,
+  })
+);
 app.use(express.json());
 app.use(cookieParser());
 
-// own  middlewares 
+// own  middlewares
 const logger = (req, res, next) => {
-    console.log('log: info', req.method, req.url);
-    next();
-}
+  console.log("log: info", req.method, req.url);
+  next();
+};
 
 const verifyToken = (req, res, next) => {
-    const token = req?.cookies?.token;
-    if (!token) {
-        return res.status(401).send({ message: 'unauthorized access' })
+  const token = req?.cookies?.token;
+  if (!token) {
+    return res.status(401).send({ message: "unauthorized access" });
+  }
+  jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, decoded) => {
+    if (err) {
+      return res.status(401).send({ message: "unauthorized access" });
     }
-    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, decoded) => {
-        if (err) {
-            return res.status(401).send({ message: 'unauthorized access' })
-        }
-        req.user = decoded;
-        next();
-    })
-}
-
+    req.user = decoded;
+    next();
+  });
+};
 
 const uri = `mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASS}@cluster0.xpzeqy0.mongodb.net/?retryWrites=true&w=majority`;
 
@@ -59,33 +59,48 @@ async function run() {
       .collection("myAssignments");
 
     // all assignments
-// get all assignment
+    // get all assignment
     app.get("/assignments", async (req, res) => {
       const cursor = assignmentCollection.find();
       const result = await cursor.toArray();
       res.send(result);
     });
-// get specific assignment
+
+    // get specific assignment
     app.get("/assignments/:id", async (req, res) => {
       const id = req.params.id;
       const query = { _id: new ObjectId(id) };
       const result = await assignmentCollection.findOne(query);
       res.send(result);
     });
-// create assignment
+
+    // create assignment
     app.post("/assignments", async (req, res) => {
       const newAssignment = req.body;
       console.log(newAssignment);
       const result = await assignmentCollection.insertOne(newAssignment);
       res.send(result);
     });
-// update assignment
+
+    // update assignment
     app.put("/assignments/:id", async (req, res) => {
       const id = req.params.id;
       const filter = { _id: new ObjectId(id) };
-      const options = { upsert: true };
-      updatedAssignment = req.body;
-      const assignment = {
+      const updatedAssignment = req.body;
+      const userEmail = updatedAssignment.email;
+      const assignment = await assignmentCollection.findOne(filter);
+
+      if (!assignment) {
+        return res.status(404).json({ message: "Assignment not found" });
+      }
+      if (assignment.email !== userEmail) {
+        return res
+          .status(403)
+          .json({
+            message: "Unauthorized: You are not the creator of this assignment",
+          });
+      }
+      const assignmentUpdate = {
         $set: {
           title: updatedAssignment.title,
           thumbnailURL: updatedAssignment.thumbnailURL,
@@ -93,25 +108,37 @@ async function run() {
           description: updatedAssignment.description,
           difficultyLevel: updatedAssignment.difficultyLevel,
           dueDate: updatedAssignment.dueDate,
+          email: userEmail,
         },
       };
+
       const result = await assignmentCollection.updateOne(
         filter,
-        assignment,
-        options
+        assignmentUpdate
       );
       res.send(result);
     });
-// delete specific assinment
+
+    
+
+    // delete specific assinment
     app.delete("/assignments/:id", async (req, res) => {
       const id = req.params.id;
       const query = { _id: new ObjectId(id) };
+      const userEmail = req.body.email;
+      const assignment = await assignmentCollection.findOne(query);
+      if (!assignment) {
+        // Assignment not found.
+        return res.status(404).json({ message: "Assignment not found" });
+      }
+      if (assignment.email !== userEmail) {
+        return res.status(403).json({
+          message: "Unauthorized: You are not the creator of this assignment",
+        });
+      }
       const result = await assignmentCollection.deleteOne(query);
       res.send(result);
     });
-
-
-
 
     // my assignment
 
@@ -134,21 +161,27 @@ async function run() {
     //   res.send(result);
     // });
 
-
-
-
     // auth
-    app.post('/jwt',logger,async(req,res)=>{
-        const user = req.body;
-        const token = jwt.sign(user,process.env.ACCESS_TOKEN_SECRET,{expiresIn: '1h'});
+    app.post("/jwt", logger, async (req, res) => {
+      const user = req.body;
+      const token = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET, {
+        expiresIn: "1h",
+      });
 
-        res.cookie('token',token,{
-            httpOnly: true,
-            secure:true,
-            sameSite:'none'
+      res
+        .cookie("token", token, {
+          httpOnly: true,
+          secure: true,
+          sameSite: "none",
         })
-        .send({success:true});
-    })
+        .send({ success: true });
+    });
+
+    app.post("/logout", async (req, res) => {
+      const user = req.body;
+      console.log("logging out", user);
+      res.clearCookie("token", { maxAge: 0 }).send({ success: true });
+    });
 
     await client.connect();
     // Send a ping to confirm a successful connection
